pyproject.toml
[project]
name = "financial-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.135.3",
    "groq>=1.1.2",
    "packaging>=26.0",
    "phidata>=2.7.10",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
    "yfinance>=1.2.1",
]
